test(button): 解决button组件单元测试问题
